requirements.txt
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Bootstrap==3.3.5.7
Flask-HTTPAuth==2.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Moment==0.5.1
flask-mongoengine==0.7.5
Flask-OpenID==1.2.5
Flask-PageDown==0.2.1
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
pymysql==0.7.2
markdown==2.6.5
bleach==1.4.2
MySQL-python==1.2.5